Switch to npm for recursive install, update all the dependencies
