+++ modules/core/content/package.json
@@ -24,8 +24,6 @@
   "devDependencies": {
     "async": "^2.6.1",
     "chai": "^4.2.0",
-    "grunt": "^1.0.3",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   },
