+++ modules/core/documentation/package.json
@@ -29,7 +29,6 @@
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
