+++ modules/core/express/package.json
@@ -17,11 +17,10 @@
   "license": "MIT",
   "dependencies": {
     "compression": "^1.7.3",
-    "express": "^4.16.3"
+    "express": "^4.16.4"
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0"
   }
 }
