+++ modules/core/io/package.json
@@ -17,7 +17,6 @@
   "license": "MIT",
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
