+++ modules/core/logging/package.json
@@ -20,12 +20,11 @@
   "license": "MIT",
   "dependencies": {
     "winston": "^3.1.0",
-    "winston-transport": "^4.2.0",
-    "winston-daily-rotate-file": "^3.3.3"
+    "winston-daily-rotate-file": "^3.5.1",
+    "winston-transport": "^4.2.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0"
   }
 }
