+++ modules/core/markdown/package.json
@@ -20,7 +20,6 @@
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0"
   }
 }
