+++ modules/core/multi-tenancy/package.json
@@ -11,7 +11,6 @@
   "license": "MIT",
   "devDependencies": {
     "chai": "^4.2.0",
-    "grunt-mocha-test": "^0.13.3",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   },
