+++ modules/core/navigation/package.json
@@ -19,6 +19,5 @@
   },
   "author": "Bertrand Le Roy",
   "license": "MIT",
-  "dependencies": {
-  }
+  "dependencies": {}
 }
