+++ modules/core/parts/package.json
@@ -16,12 +16,11 @@
   "author": "Bertrand Le Roy",
   "license": "MIT",
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
-    "mocha": "^5.2.0",
-    "async": "^2.6.1"
+    "async": "^2.6.1",
+    "chai": "^4.2.0",
+    "mocha": "^5.2.0"
   },
   "dependencies": {
-    "htmlencode": "^0.0.4"
+    "htmlencode": "0.0.4"
   }
 }
