+++ modules/core/routing/package.json
@@ -33,8 +33,7 @@
   "author": "Bertrand Le Roy",
   "license": "MIT",
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
+    "chai": "^4.2.0",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
