+++ modules/core/tokens/package.json
@@ -23,8 +23,7 @@
     "markup-js": "^1.5.21"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
+    "chai": "^4.2.0",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
