+++ modules/core/ui/package.json
@@ -44,8 +44,7 @@
   "author": "Bertrand Le Roy",
   "license": "MIT",
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
+    "chai": "^4.2.0",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   },
@@ -56,6 +55,6 @@
     "flasync": "^0.2.3",
     "htmlencode": "0.0.4",
     "js-object-pretty-print": "^0.3.0",
-    "luxon": "^1.4.3"
+    "luxon": "^1.6.2"
   }
 }
