+++ modules/core/widgets/package.json
@@ -25,8 +25,7 @@
     "static-eval": "^2.0.0"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
+    "chai": "^4.2.0",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
