+++ modules/db/couch/package.json
@@ -24,8 +24,7 @@
     "test": "./test/index"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "grunt-mocha-test": "^0.13.3",
+    "chai": "^4.2.0",
     "mocha": "^5.2.0",
     "proxyquire": "^2.1.0"
   }
