+++ package.json
@@ -22,15 +22,13 @@
     "email": "bertrand@decentcms.com",
     "url": "http://weblogs.asp.net/bleroy"
   },
+  "scripts": {
+    "postinstall": "node ./lib/install-decent-modules.js -- install"
+  },
   "dependencies": {},
   "devDependencies": {
     "async": "^2.6.1",
     "chai": "^4.2.0",
-    "grunt": "^1.0.3",
-    "grunt-cli": "^1.3.1",
-    "grunt-contrib-imagemin": "gruntjs/grunt-contrib-imagemin",
-    "grunt-mocha-test": "^0.13.3",
-    "grunt-run-grunt": "^1.0.1",
     "mocha": "^5.2.0"
   }
 }
